+++ CHANGELOG.md
@@ -13,3 +13,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `scroll` and `showContentDividers` props.
 - Dismiss when press `Esc` key.
 - Basic support to animations + `Fade` animation component.
+- Create the `modal-content` and `modal-header` blocks.